# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization API",
        "title": "Auth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/pdf/crop": {
            "post": {
                "description": "PDF sahifalarini belgilangan tomonlardan qirqadi (top, bottom, left, right)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Edit"
                ],
                "summary": "Crop PDF pages",
                "parameters": [
                    {
                        "description": "Crop request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CropPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/crop/{id}": {
            "get": {
                "description": "Crop PDF jarayonining natijasini olish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Edit"
                ],
                "summary": "Get Crop Job info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CropPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/pdf/extract": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-extract"
                ],
                "summary": "Create extract job",
                "parameters": [
                    {
                        "description": "extract request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExtractPagesRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/extract/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-extract"
                ],
                "summary": "Get extract job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "extract job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExtractJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CropPDFJob": {
            "type": "object",
            "properties": {
                "bottom": {
                    "description": "Pastki chegara",
                    "type": "integer"
                },
                "box": {
                    "description": "Box turi",
                    "type": "string"
                },
                "created_at": {
                    "description": "Yaratilgan vaqt",
                    "type": "string"
                },
                "id": {
                    "description": "Ish IDsi",
                    "type": "string"
                },
                "input_file_id": {
                    "description": "Kirish fayl IDsi",
                    "type": "string"
                },
                "input_file_ids": {
                    "description": "Fayl IDlari ro'yxati (bir nechta fayl uchun)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "left": {
                    "description": "Chap chegara",
                    "type": "integer"
                },
                "output_file_id": {
                    "description": "Chiqish fayli IDsi",
                    "type": "string"
                },
                "pages": {
                    "description": "Sahifalar",
                    "type": "string"
                },
                "right": {
                    "description": "O'ng chegara",
                    "type": "integer"
                },
                "status": {
                    "description": "Ish holati (masalan, \"pending\", \"done\")",
                    "type": "string"
                },
                "top": {
                    "description": "Yuqori chegara",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi IDsi (guest uchun nil bo'lishi mumkin)",
                    "type": "string"
                }
            }
        },
        "models.CropPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "bottom": {
                    "type": "integer"
                },
                "box": {
                    "description": "optional: mediabox, cropbox, etc",
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "left": {
                    "type": "integer"
                },
                "pages": {
                    "description": "optional: qaysi sahifalarni crop qilish",
                    "type": "string"
                },
                "right": {
                    "type": "integer"
                },
                "top": {
                    "type": "integer"
                }
            }
        },
...
        "models.ExtractJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "pages_to_extract": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.ExtractPagesRequest": {
            "type": "object",
            "required": [
                "input_file_id",
                "page_ranges"
            ],
            "properties": {
                "input_file_id": {
                    "type": "string"
                },
                "page_ranges": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}